+++ djangorestframework_camel_case/__init__.py
@@ -1,9 +1,3 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-
-__author__ = 'Vitaly Babiy'
-__email__ = 'vbabiy86@gmail.com'
-__version__ = '0.1.2'
-
-from render import CamelCaseJSONRenderer
-from parser import CamelCaseJSONParser
+__author__ = "Vitaly Babiy"
+__email__ = "vbabiy86@gmail.com"
+__version__ = "1.4.2"